Update and rename application-template.md to WorldBox.md
I still need to lock down the finer details for the Milestones (still in discussion with Touch Foundry). These will be updated in the next few days. However the general milestones for the project as a whole can be found on the www.worldbox.tv website (soft launched earlier today).